Implement lambda-calculus in types
Looking at the solution of @vasyoid, I was interested in the method
he chose for computing values, using not implicit values and
logic-like programming but total computations on type synonyms. I
decided to write a lambda calculator in it, hoping to maybe use it
to write high-level programs without all the boilerplate, using
only the rich mechanics of Church encodings.

This method of computations doesn't work as well as I'd hoped. It
doesn't crash the compiler as the current one does, but it takes
forever to evaluate something as simple as 10 * 10 in plain natural
numbers. So maybe this approach isn't easier after all.

I'll keep the file just as a curiosity.